Refactoring async-generic event subscription
