+++ crates/relayer-components-extra/src/components/extra/closures/relay/auto_relayer.rs
@@ -7,6 +7,8 @@ use ibc_relayer_components::logger::traits::has_logger::{HasLogger, HasLoggerTyp
 use ibc_relayer_components::logger::traits::level::HasBaseLogLevels;
 use ibc_relayer_components::relay::traits::chains::HasRelayChains;
 use ibc_relayer_components::runtime::traits::runtime::HasRuntime;
+use ibc_relayer_components::runtime::traits::stream::CanMapStream;
+use ibc_relayer_components::runtime::traits::subscription::HasSubscriptionType;
 use ibc_relayer_components::runtime::traits::task::CanRunConcurrentTasks;
 
 use crate::components::extra::closures::relay::event_relayer::UseExtraEventRelayer;
@@ -27,18 +29,20 @@ where
         + HasComponents<Components = ExtraRelayComponents<BaseRelayComponents>>,
     Relay::SrcChain: HasRuntime
         + HasChainId
-        + HasEventSubscription
         + HasLoggerType<Logger = Relay::Logger>
-        + CanLogChainEvent,
+        + CanLogChainEvent
+        + HasEventSubscription,
     Relay::DstChain: HasRuntime
         + HasChainId
-        + HasEventSubscription
         + HasLoggerType<Logger = Relay::Logger>
-        + CanLogChainEvent,
+        + CanLogChainEvent
+        + HasEventSubscription,
     Relay::Runtime: HasSpawner + CanRunConcurrentTasks,
     Relay::Logger: HasBaseLogLevels,
-    <Relay::SrcChain as HasRuntime>::Runtime: HasSpawner,
-    <Relay::DstChain as HasRuntime>::Runtime: HasSpawner,
+    <Relay::SrcChain as HasRuntime>::Runtime:
+        HasSubscriptionType + CanRunConcurrentTasks + CanMapStream,
+    <Relay::DstChain as HasRuntime>::Runtime:
+        HasSubscriptionType + CanRunConcurrentTasks + CanMapStream,
     BaseRelayComponents: Async,
 {
 }
