+++ crates/relayer-components-extra/src/relay/components/auto_relayers/mod.rs
@@ -1,2 +1,1 @@
-pub mod parallel_event;
 pub mod parallel_two_way;
