--- crates/relayer-components-extra/src/runtime/impls/subscription/mod.rs
@@ -1,2 +1,0 @@
-pub mod multiplex;
-pub mod stream;
