+++ crates/relayer-components/src/components/default/closures/relay/auto_relayer.rs
@@ -7,6 +7,8 @@ use crate::components::default::relay::DefaultRelayComponents;
 use crate::core::traits::run::CanRun;
 use crate::relay::traits::chains::HasRelayChains;
 use crate::runtime::traits::runtime::HasRuntime;
+use crate::runtime::traits::stream::CanMapStream;
+use crate::runtime::traits::subscription::HasSubscriptionType;
 use crate::runtime::traits::task::CanRunConcurrentTasks;
 
 pub trait CanUseDefaultAutoRelayer: UseDefaultAutoRelayer {}
@@ -23,6 +25,10 @@ where
     Relay::SrcChain: HasEventSubscription,
     Relay::DstChain: HasEventSubscription,
     Relay::Runtime: CanRunConcurrentTasks,
+    <Relay::SrcChain as HasRuntime>::Runtime:
+        HasSubscriptionType + CanRunConcurrentTasks + CanMapStream,
+    <Relay::DstChain as HasRuntime>::Runtime:
+        HasSubscriptionType + CanRunConcurrentTasks + CanMapStream,
     BaseRelayComponents: Async,
 {
 }
