+++ crates/relayer-components/src/components/default/relay.rs
@@ -4,7 +4,7 @@ use cgp_core::delegate_component;
 
 use crate::core::traits::run::RunnerComponent;
 use crate::relay::components::auto_relayers::bidirectional::BidirectionalRelayer;
-use crate::relay::components::auto_relayers::concurrent_event::ConcurrentEventSubscriptionRelayer;
+use crate::relay::components::auto_relayers::event::RelayEvents;
 use crate::relay::components::create_client::CreateClientWithChains;
 use crate::relay::components::event_relayers::packet_event::PacketEventRelayer;
 use crate::relay::components::message_senders::chain_sender::SendIbcMessagesToChain;
@@ -112,7 +112,7 @@ delegate_component!(
 delegate_component!(
     AutoRelayerComponent,
     DefaultRelayComponents<BaseComponents>,
-    ConcurrentEventSubscriptionRelayer,
+    RelayEvents,
 );
 
 delegate_component!(
