+++ crates/relayer-components/src/relay/components/auto_relayers/mod.rs
@@ -1,5 +1,3 @@
 pub mod bidirectional;
-pub mod concurrent_event;
 pub mod concurrent_two_way;
 pub mod event;
-pub mod sequential_event;
