--- crates/relayer-components/src/runtime/impls/mod.rs
@@ -1,1 +1,0 @@
-pub mod subscription;
