+++ crates/relayer-components/src/runtime/mod.rs
@@ -2,6 +2,5 @@
    Constructs for the runtime context.
 */
 
-pub mod impls;
 pub mod traits;
 pub mod types;
