+++ crates/relayer-runtime/src/impls/runtime/stream.rs
@@ -2,10 +2,24 @@ use core::pin::Pin;
 
 use cgp_core::traits::Async;
 use futures::stream::Stream;
-use ibc_relayer_components::runtime::traits::stream::HasStreamType;
+use futures::StreamExt;
+use ibc_relayer_components::runtime::traits::stream::{CanMapStream, HasStreamType};
 
 use crate::types::runtime::TokioRuntimeContext;
 
 impl HasStreamType for TokioRuntimeContext {
     type Stream<Item: Async> = Pin<Box<dyn Stream<Item = Item> + Send + Sync + 'static>>;
+}
+
+impl CanMapStream for TokioRuntimeContext {
+    fn map_stream<T, U, M>(stream: Self::Stream<T>, mapper: M) -> Self::Stream<U>
+    where
+        T: Async,
+        U: Async,
+        M: Fn(T) -> U,
+    {
+        let mapped = stream.map(mapper);
+
+        Box::pin(mapped)
+    }
 }
