+++ crates/relayer-runtime/src/lib.rs
@@ -1,4 +1,5 @@
 extern crate alloc;
 
 pub mod impls;
+pub mod traits;
 pub mod types;
